chore(payment): DRY payment index in controller and test
